CarbCount/requirements.txt
django
django-extensions
fatsecret
pylint
django-rest-framework
django-rest-auth
psycopg2
datetimerange